mpi: rename mpi_to_llvm to lower_mpi in code
