+++ tests/dialects/test_mpi_lowering.py
@@ -1,10 +1,10 @@
 from xdsl.dialects import mpi, func, llvm, builtin
 from xdsl.ir import Operation, Attribute, OpResult, SSAValue
 from xdsl.irdl import irdl_op_definition, VarOpResult
-from xdsl.transforms import mpi_to_llvm
+from xdsl.transforms import lower_mpi
 from xdsl.dialects.builtin import i32
 
-info = mpi_to_llvm.MpiLibraryInfo()
+info = lower_mpi.MpiLibraryInfo()
 
 
 def extract_func_call(ops: list[Operation],
@@ -50,7 +50,7 @@ class CreateTestValsOp(Operation):
 
 
 def test_lower_mpi_init():
-    ops, result = mpi_to_llvm.LowerMpiInit(info).lower(mpi.Init.build())
+    ops, result = lower_mpi.LowerMpiInit(info).lower(mpi.Init.build())
 
     assert len(result) == 0
     assert len(ops) == 2
@@ -65,8 +65,7 @@ def test_lower_mpi_init():
 
 
 def test_lower_mpi_finalize():
-    ops, result = mpi_to_llvm.LowerMpiFinalize(info).lower(
-        mpi.Finalize.build())
+    ops, result = lower_mpi.LowerMpiFinalize(info).lower(mpi.Finalize.build())
 
     assert len(result) == 0
     assert len(ops) == 1
@@ -81,7 +80,7 @@ def test_lower_mpi_finalize():
 def test_lower_mpi_wait_no_status():
     request, = CreateTestValsOp.get(mpi.RequestType()).results
 
-    ops, result = mpi_to_llvm.LowerMpiWait(info).lower(mpi.Wait.get(request))
+    ops, result = lower_mpi.LowerMpiWait(info).lower(mpi.Wait.get(request))
 
     assert len(result) == 0
     call = extract_func_call(ops)
@@ -93,7 +92,7 @@ def test_lower_mpi_wait_no_status():
 def test_lower_mpi_wait_with_status():
     request, = CreateTestValsOp.get(mpi.RequestType()).results
 
-    ops, result = mpi_to_llvm.LowerMpiWait(info).lower(
+    ops, result = lower_mpi.LowerMpiWait(info).lower(
         mpi.Wait.get(request, ignore_status=False))
 
     assert len(result) == 1
@@ -108,7 +107,7 @@ def test_lower_mpi_wait_with_status():
 
 
 def test_lower_mpi_comm_rank():
-    ops, result = mpi_to_llvm.LowerMpiCommRank(info).lower(mpi.CommRank.get())
+    ops, result = lower_mpi.LowerMpiCommRank(info).lower(mpi.CommRank.get())
 
     assert len(result) == 1
     assert result[0] is not None
@@ -128,7 +127,7 @@ def test_lower_mpi_send():
         mpi.MemRefType.from_element_type_and_shape(builtin.f64, [32, 32, 32]),
         i32).results
 
-    ops, result = mpi_to_llvm.LowerMpiSend(info).lower(
+    ops, result = lower_mpi.LowerMpiSend(info).lower(
         mpi.Send.get(buff, dest, 1))
     """
     Check for function with signature like:
@@ -150,7 +149,7 @@ def test_lower_mpi_isend():
         mpi.MemRefType.from_element_type_and_shape(builtin.f64, [32, 32, 32]),
         i32).results
 
-    ops, result = mpi_to_llvm.LowerMpiISend(info).lower(
+    ops, result = lower_mpi.LowerMpiISend(info).lower(
         mpi.ISend.get(buff, dest, 1))
     """
     Check for function with signature like:
@@ -175,7 +174,7 @@ def test_lower_mpi_recv_no_status():
              MPI_Comm comm, MPI_Status *status)
     """
 
-    ops, result = mpi_to_llvm.LowerMpiRecv(info).lower(
+    ops, result = lower_mpi.LowerMpiRecv(info).lower(
         mpi.Recv.get(source, buff, tag=3, ignore_status=True))
 
     assert len(result) == 0
@@ -194,7 +193,7 @@ def test_lower_mpi_recv_with_status():
              MPI_Comm comm, MPI_Status *status)
     """
 
-    ops, result = mpi_to_llvm.LowerMpiRecv(info).lower(
+    ops, result = lower_mpi.LowerMpiRecv(info).lower(
         mpi.Recv.get(source, buff, tag=3, ignore_status=False))
 
     assert len(result) == 1
@@ -213,7 +212,7 @@ def test_lower_mpi_irecv():
             int source, int tag, MPI_Comm comm, MPI_Request *request)
     """
 
-    ops, result = mpi_to_llvm.LowerMpiIRecv(info).lower(
+    ops, result = lower_mpi.LowerMpiIRecv(info).lower(
         mpi.IRecv.get(source, buff, tag=3))
 
     assert len(result) == 1
@@ -228,7 +227,7 @@ def test_mpi_type_conversion():
     Test that each builtin datatype is correctly mapped to an MPI datatype
 
     """
-    info = mpi_to_llvm.MpiLibraryInfo(
+    info = lower_mpi.MpiLibraryInfo(
         MPI_UNSIGNED_CHAR=1,
         MPI_UNSIGNED_SHORT=2,
         MPI_UNSIGNED=3,
@@ -239,7 +238,7 @@ def test_mpi_type_conversion():
         MPI_LONG_LONG_INT=8,
     )
 
-    lowering = mpi_to_llvm.LowerMpiRecv(info)
+    lowering = lower_mpi.LowerMpiRecv(info)
 
     from xdsl.dialects.builtin import f64, f32, IntegerType, i32, i64, Signedness
     u64 = IntegerType(64, Signedness.UNSIGNED)
